Profits Money amounts are now decimals ProfitOfPeriods: + new RoundedNetIncome property * money amounts are now decimal
ReceiptsViewModel:
* money amounts are now decimal

+++ libre-pensador-api/libre-pensador-api/CRUD/ProfitPerPeriodsService.cs
@@ -78,8 +78,8 @@ namespace libre_pensador_api.CRUD
                 }
                 else
                 {
-                    profitPerSubPeriod[currentSubPeriodIndex].IncomeBeforeTaxes -= (double)expensesEnumerator.Current.AmountSpent;
-                    profitPerSubPeriod[currentSubPeriodIndex].NetIncome -= (double)expensesEnumerator.Current.AmountSpent;
+                    profitPerSubPeriod[currentSubPeriodIndex].IncomeBeforeTaxes -= expensesEnumerator.Current.AmountSpent;
+                    profitPerSubPeriod[currentSubPeriodIndex].NetIncome -= expensesEnumerator.Current.AmountSpent;
                     expensesEnumerator.MoveNext();
                 }
             }
